Change requests_cache implementation.
This makes for a more minor change to the way the mtgjson library works. You
get an optional parameter to the from_url() method, and the mtgjson library
doesn't need to depend on requests_cache... though I'm not sure how I
express in the setup.py that requests_cache is now only needed for running
the tests.

I've also removed the download_set_file() in the tests. I think this fixture
was trying to save downloads by downloading the MTGJSON All Sets file once,
saving it to a file, and using the file subsequently. If that is what was
intended, it's not what is happening. What is happening is each test is run
twice, once with a download and once from the downloaded file. Which makes
for a slow execution of the tests.

So I've added use of reqquest_cache to the tests, to make it a bit faster to
run the tests.  I've included in this commit some changes to that the
un-cached code path is only run if you pass '--un-cached' command option.